# interpreter/go: replace Rust with native Go symbolization