--- a/.gitignore
+++ b/.gitignore
@@ -7,6 +7,3 @@ ebpf-profiler
 ci-kernels
 # Ignore target directory
 target/*
-# But not these specific paths
-!target/x86_64-unknown-linux-musl/release/libsymblib_capi.a
-!target/aarch64-unknown-linux-musl/release/libsymblib_capi.a