… interpreter/go/go_amd64.go
… interpreter/go/go_arm64.go
… tools/coredump/testdata/arm64/go.symbhack.readheader.json