This is a follow up PR show casing the use of https://github.com/open-telemetry/opentelemetry-ebpf-profiler/pull/455.

A side effect of this change is the reduced complexity of compilation, less cgo calls, reduction of binary size.

Fixes https://github.com/open-telemetry/opentelemetry-ebpf-profiler/issues/457
Fixes https://github.com/open-telemetry/opentelemetry-ebpf-profiler/issues/512

Benchmark results of [current main](https://github.com/open-telemetry/opentelemetry-ebpf-profiler/commit/1f7906203b301313c73889b3abcfe563c0847217) vs this branch:
```
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/ebpf-profiler/interpreter/go
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
          │     old.txt     │               new.txt               │
          │     sec/op      │   sec/op     vs base                │
Golang-20   6375752.5n ± 6%   969.3n ± 2%  -99.98% (p=0.000 n=10)

          │  old.txt   │              new.txt               │
          │    B/op    │    B/op     vs base                │
Golang-20   848.5 ± 1%   632.0 ± 0%  -25.52% (p=0.000 n=10)

          │  old.txt   │              new.txt               │
          │ allocs/op  │ allocs/op   vs base                │
Golang-20   24.00 ± 0%   17.00 ± 0%  -29.17% (p=0.000 n=10)
```

I think it is important to state, that the Rust based symbolization is not the isse but the overhead introduced by cgo calling the Rust implementation in Go.